feat(ui): add Rate submenu to track context menu; remove dead ContextMenus.swift
The right-click context menu was missing the Rate submenu that exists in
the global Track menu. Wire it end-to-end:

- TrackContextMenuActions: add rate: ([Track], Int) -> Void property
- LibraryViewModel+Rating: add setRating(stars:for:) that accepts explicit
  tracks (bypasses selection state, used by the AppKit coordinator which
  already has the target tracks in hand)
- TracksView+Actions: wire rate closure to lib.setRating(stars:for:)
- TrackTableCoordinator: add addRateItem() — Rate submenu with
  None/★☆☆☆☆/…/★★★★★ entries after the Love/Unlove item

Dead code removal:
- Delete ContextMenus.swift (TrackActions struct and TrackContextMenu SwiftUI
  view were never instantiated; helper extensions revealInFinder/copyAsTSV
  were also unused)

Housekeeping:
- Extract ActionMenuItem, ShuffleCheckCell, TrackDiffableDataSource from
  TrackTableCoordinator.swift into TrackTableHelpers.swift to keep the
  coordinator under the 500-line SwiftLint limit
